#1 Refactoring chainer to combinator. Still need …
…to add docs and tests before actually publishing anything. Most
likely build process tweaks are needed as well, to do .mjs renaming.
