+++ data/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ty-ras/data",
-  "version": "1.0.0",
+  "version": "0.0.1",
   "author": {
     "name": "Stanislav Muhametsin",
     "email": "346799+stazz@users.noreply.github.com",
@@ -11,7 +11,21 @@
     "type": "git",
     "url": "https://github.com/ty-ras/data"
   },
-  "type": "module",
+  "files": [
+    "./src",
+    "./dist",
+    "README.md"
+  ],
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "require": "./dist/index.js",
+      "import": "./dist/index.mjs",
+      "types": "./dist/index.d.ts"
+    }
+  },
   "scripts": {
     "build": "yarn run tsc && yarn run lint",
     "tsc": "tsc",
