+++ data/src/values.d.ts
@@ -1,0 +1,3 @@
+export type HeaderValue = string | Array<string> | undefined;
+
+export type QueryValue = string | Array<string> | undefined;
